Conflicts:
	MakeDefaults/Lint/marte_flint_files.lnt
	Source/Core/BareMetal/L0Types/Architecture/x86_cl/CompilerTypes.h
	Source/Core/BareMetal/L0Types/CString.h
	Source/Core/BareMetal/L0Types/LinkedListHolder.cpp
	Source/Core/BareMetal/L0Types/LinkedListHolderT.h
	Source/Core/BareMetal/L0Types/SearchFilterT.h
	Source/Core/BareMetal/L0Types/ZeroTerminatedArray.h
	Source/Core/BareMetal/L1Portability/Architecture/x86_cl/HighResolutionTimerA.h
	Source/Core/BareMetal/L1Portability/Architecture/x86_gcc/HighResolutionTimerA.h
	Source/Core/BareMetal/L1Portability/Environment/Linux/HighResolutionTimer.h
	Source/Core/BareMetal/L1Portability/Environment/Linux/HighResolutionTimerCalibrator.cpp
	Source/Core/BareMetal/L1Portability/Environment/Linux/HighResolutionTimerCalibrator.h
	Source/Core/BareMetal/L1Portability/Environment/Linux/Sleep.cpp
	Source/Core/BareMetal/L1Portability/Environment/Windows/HighResolutionTimerCalibrator.cpp
	Source/Core/BareMetal/L1Portability/Environment/Windows/HighResolutionTimerCalibrator.h
	Source/Core/BareMetal/L1Portability/FastPollingEventSem.cpp
	Source/Core/BareMetal/L1Portability/FastPollingEventSem.h
	Source/Core/BareMetal/L1Portability/FastPollingMutexSem.cpp
	Source/Core/BareMetal/L1Portability/FastPollingMutexSem.h
	Source/Core/BareMetal/L1Portability/GlobalObjectsDatabase.cpp
	Source/Core/BareMetal/L1Portability/HighResolutionTimer.h
	Source/Core/BareMetal/L1Portability/Makefile.inc
	Source/Core/BareMetal/L1Portability/MemoryArea.cpp
	Source/Core/BareMetal/L1Portability/MemoryArea.h
	Source/Core/BareMetal/L1Portability/Sleep.h
	Source/Core/BareMetal/L1Portability/StaticList.h
	Source/Core/BareMetal/L2Objects/ClassMethodInterfaceMapper.h
	Source/Core/BareMetal/L2Objects/ClassMethodsRegistryItem.cpp
	Source/Core/BareMetal/L2Objects/ClassMethodsRegistryItem.h
	Source/Core/BareMetal/L2Objects/ClassProperties.cpp
	Source/Core/BareMetal/L2Objects/ClassProperties.h
	Source/Core/BareMetal/L2Objects/ClassRegistryDatabase.cpp
	Source/Core/BareMetal/L2Objects/ClassRegistryDatabase.h
	Source/Core/BareMetal/L2Objects/ClassRegistryItem.cpp
	Source/Core/BareMetal/L2Objects/ClassRegistryItem.h
	Source/Core/BareMetal/L2Objects/Introspection.h
	Source/Core/BareMetal/L2Objects/Makefile.inc
	Source/Core/BareMetal/L2Objects/Object.cpp
	Source/Core/BareMetal/L2Objects/Object.h
	Source/Core/BareMetal/L2Objects/ObjectBuilder.h
	Source/Core/BareMetal/L2Objects/ObjectRegistryDatabase.cpp
	Source/Core/BareMetal/L2Objects/ObjectRegistryDatabase.h
	Source/Core/BareMetal/L2Objects/Reference.cpp
	Source/Core/BareMetal/L2Objects/Reference.h
	Source/Core/BareMetal/L2Objects/ReferenceContainer.cpp
	Source/Core/BareMetal/L2Objects/ReferenceContainer.h
	Source/Core/BareMetal/L2Objects/ReferenceContainerFilterObjectName.cpp
	Source/Core/BareMetal/L2Objects/ReferenceT.h
	Source/Core/BareMetal/L2Objects/StructuredDataI.h
	Source/Core/BareMetal/L3Streams/IOBuffer.cpp
	Source/Core/BareMetal/L4Messages/.gitignore
	Source/Core/BareMetal/L4Messages/Message.cpp
	Source/Core/BareMetal/L4Messages/Message.h
	Source/Core/BareMetal/L4Messages/MessageI.cpp
	Source/Core/BareMetal/L4Messages/MessageI.h
	Source/Core/BareMetal/Makefile.cov
	Source/Core/BareMetal/Makefile.inc
	Source/Core/BareMetal/Makefile.msc
	Source/Core/Scheduler/L1Portability/Makefile.inc
	Source/Core/Scheduler/L1Portability/Makefile.msc
	Test/Core/BareMetal/L1Portability/MemoryAreaTest.cpp
	Test/Core/BareMetal/L1Portability/MemoryAreaTest.h
	Test/Core/BareMetal/L2Objects/AnyTypeTest.cpp
	Test/Core/BareMetal/L2Objects/ClassRegistryDatabaseTest.cpp
	Test/Core/BareMetal/L2Objects/Makefile.inc
	Test/Core/BareMetal/L2Objects/ObjectRegistryDatabaseTest.cpp
	Test/Core/BareMetal/L2Objects/ObjectRegistryDatabaseTest.h
	Test/Core/BareMetal/L2Objects/ObjectTest.cpp
	Test/Core/BareMetal/L2Objects/ObjectTest.h
	Test/Core/BareMetal/L2Objects/ObjectTestHelper.h
	Test/Core/BareMetal/L2Objects/ReferenceContainerTest.cpp
	Test/Core/BareMetal/L2Objects/ReferenceContainerTest.h
	Test/Core/BareMetal/L2Objects/TestObjectHelper3.cpp
	Test/Core/BareMetal/L2Objects/TestObjectHelper3.h
	Test/Core/BareMetal/L4Configuration/IntrospectionTestHelper.cpp
	Test/Core/BareMetal/Makefile.cov
	Test/Core/BareMetal/Makefile.inc
	Test/GTest/.gitignore
	Test/GTest/BareMetal/L0Types/.gitignore
	Test/GTest/BareMetal/L1Portability/.gitignore
	Test/GTest/BareMetal/L1Portability/FastPollingEventSemGTest.cpp
	Test/GTest/BareMetal/L1Portability/MemoryAreaGTest.cpp
	Test/GTest/BareMetal/L2Objects/.gitignore
	Test/GTest/BareMetal/L2Objects/ObjectGTest.cpp
	Test/GTest/BareMetal/L2Objects/ObjectRegistryDatabaseGTest.cpp
	Test/GTest/BareMetal/L2Objects/ReferenceContainerFilterObjectNameGTest.cpp
	Test/GTest/BareMetal/L2Objects/ReferenceContainerGTest.cpp
	Test/GTest/BareMetal/L3Streams/.gitignore
	Test/GTest/BareMetal/L4Configuration/.gitignore
	Test/GTest/BareMetal/Makefile.cov
	Test/GTest/BareMetal/Makefile.inc
	Test/GTest/FileSystem/L1Portability/.gitignore
	Test/GTest/Scheduler/L1Portability/.gitignore

+++ Source/Core/BareMetal/L1Portability/Environment/Linux/HighResolutionTimerCalibrator.cpp
@@ -55,7 +55,7 @@ namespace MARTe {
 HighResolutionTimerCalibrator calibratedHighResolutionTimer;
 
 HighResolutionTimerCalibrator::HighResolutionTimerCalibrator() {
-    const size_t LINUX_CPUINFO_BUFFER_SIZE = 1023u;
+    const uint64 LINUX_CPUINFO_BUFFER_SIZE = 1023u;
     initialTicks = HighResolutionTimer::Counter();
     frequency = 0u;
     period = 0.;
@@ -73,7 +73,7 @@ HighResolutionTimerCalibrator::HighResolutionTimerCalibrator() {
         FILE *f = fopen("/proc/cpuinfo", "r");
         size_t size = LINUX_CPUINFO_BUFFER_SIZE;
         if (f != NULL) {
-            size = fread(&buffer[0], static_cast<size_t>(size), static_cast<size_t>(1u), f);
+            size = fread(&buffer[0], size, static_cast<size_t>(1u), f);
             fclose(f);
         }
         else {
@@ -105,7 +105,7 @@ HighResolutionTimerCalibrator::HighResolutionTimerCalibrator() {
 
 bool HighResolutionTimerCalibrator::GetTimeStamp(TimeStamp &timeStamp) const {
 
-    int64 ticksFromStart = HighResolutionTimer::Counter() - initialTicks;
+    uint64 ticksFromStart = HighResolutionTimer::Counter() - initialTicks;
 
     //Use HRT
     float64 secondsFromStart = static_cast<float64>(ticksFromStart) * period;
